backend/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
pymongo
redis
requests
PyJWT
python-dotenv